Move IQuadDestination to separate file
