+++ packages/bus-rdf-update-quads/index.ts
@@ -1,2 +1,3 @@
 export * from './lib/ActorRdfUpdateQuads';
 export * from './lib/ActorRdfUpdateQuadsDestination';
+export * from './lib/IQuadDestination';
